Added tests for CAS Data export
