+++ custom/icds_reports/views.py
@@ -44,6 +44,7 @@ from custom.icds_reports.const import LocationTypes, BHD_ROLE, ICDS_SUPPORT_EMAI
     BENEFICIARY_LIST_EXPORT, ISSNIP_MONTHLY_REGISTER_PDF, AWW_INCENTIVE_REPORT, INDIA_TIMEZONE
 from custom.icds_reports.models.helper import IcdsFile
 from custom.icds_reports.models.views import AwcLocationMonths
+from custom.icds_reports.queries import get_cas_data_blob_file
 from custom.icds_reports.reports.adhaar import get_adhaar_data_chart, get_adhaar_data_map, get_adhaar_sector_data
 from custom.icds_reports.reports.adolescent_girls import get_adolescent_girls_data_map, \
     get_adolescent_girls_sector_data, get_adolescent_girls_data_chart
@@ -1725,20 +1726,13 @@ class DishaAPIView(View):
 class CasDataExport(View):
     def post(self, request, *args, **kwargs):
 
-        indicators = ['', 'child_health_monthly', 'ccs_record_monthly', 'agg_awc']
-
         data_type = int(request.POST.get('indicator', None))
         state_id = request.POST.get('location', None)
         month = request.POST.get('month', None)
         year = request.POST.get('year', None)
-        date = '{}-{}-01'.format(year, month)
-        blob_id = "{}-{}-{}".format(
-            indicators[data_type],
-            state_id,
-            date
-        )
+        selected_date = date(year, month, 1).strftime('%Y-%m-%d')
 
-        sync = IcdsFile.objects.get(blob_id=blob_id)
+        sync = get_cas_data_blob_file(data_type, state_id, selected_date)
 
         zip_name = 'cas_data_%s' % sync.file_added.strftime('%Y-%m-%d')
         try:
